+++ src/main/java/kitchenpos/menu/application/MenuGroupService.java
@@ -3,7 +3,7 @@ package kitchenpos.menu.application;
 import java.util.List;
 import kitchenpos.menu.domain.MenuGroupRepository;
 import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.dto.MenuGroupCreateRequest;
+import kitchenpos.menu.application.dto.MenuGroupCreateRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
