+++ src/main/java/kitchenpos/menu/ui/MenuGroupRestController.java
@@ -3,7 +3,7 @@ package kitchenpos.menu.ui;
 import java.net.URI;
 import java.util.List;
 import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.dto.MenuGroupCreateRequest;
+import kitchenpos.menu.application.dto.MenuGroupCreateRequest;
 import kitchenpos.menu.application.MenuGroupService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
