+++ src/main/java/kitchenpos/order/application/dto/OrderLineItemCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.dto;
+package kitchenpos.order.application.dto;
 
 public class OrderLineItemCreateRequest {
     private Long menuId;
