+++ src/main/java/kitchenpos/order/ui/OrderRestController.java
@@ -2,8 +2,8 @@ package kitchenpos.order.ui;
 
 import kitchenpos.order.application.OrderService;
 import kitchenpos.order.domain.Order;
-import kitchenpos.order.dto.OrderCreateRequest;
-import kitchenpos.order.dto.OrderUpdateRequest;
+import kitchenpos.order.application.dto.OrderCreateRequest;
+import kitchenpos.order.application.dto.OrderUpdateRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
