+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -3,7 +3,7 @@ package kitchenpos.product.application;
 import java.util.List;
 import kitchenpos.product.domain.ProductRepository;
 import kitchenpos.product.domain.Product;
-import kitchenpos.product.dto.ProductCreateRequest;
+import kitchenpos.product.application.dto.ProductCreateRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
