+++ src/main/java/kitchenpos/table/application/dto/OrderTableCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.dto;
+package kitchenpos.table.application.dto;
 
 public class OrderTableCreateRequest {
     private int numberOfGuests;
