+++ src/test/java/kitchenpos/menu/MenuServiceTest.java
@@ -10,8 +10,8 @@ import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 import kitchenpos.support.ServiceTest;
-import kitchenpos.menu.dto.MenuCreateRequest;
-import kitchenpos.menu.dto.MenuProductCreateRequest;
+import kitchenpos.menu.application.dto.MenuCreateRequest;
+import kitchenpos.menu.application.dto.MenuProductCreateRequest;
 import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.domain.MenuGroupRepository;
 import kitchenpos.menu.domain.MenuProductRepository;
