+++ src/test/java/kitchenpos/order/OrderServiceTest.java
@@ -13,9 +13,9 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import kitchenpos.support.ServiceTest;
 import kitchenpos.order.application.OrderService;
-import kitchenpos.order.dto.OrderCreateRequest;
-import kitchenpos.order.dto.OrderLineItemCreateRequest;
-import kitchenpos.order.dto.OrderUpdateRequest;
+import kitchenpos.order.application.dto.OrderCreateRequest;
+import kitchenpos.order.application.dto.OrderLineItemCreateRequest;
+import kitchenpos.order.application.dto.OrderUpdateRequest;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.table.domain.OrderTableRepository;
 import kitchenpos.order.domain.Order;
